+++ frontend/src/app/account/messages/create/page.js
@@ -1,5 +1,5 @@
-import FormCreateAdvert from "@/app/_ui/form/FormCreateAdvert";
 import {auth} from "@/auth";
+import AdvertForm from "@/app/_ui/form/AdvertForm";
 
 export const metadata = {
     title: 'Новая заявка'
@@ -9,8 +9,9 @@ export default async function Page(){
     return (
     <div className="flex flex-col gap-6 mt-1 items-center ">
                 <h2 className="text-3xl font-semibold">Заявка на сбыт отходов</h2>
-        <FormCreateAdvert userData={session?.user}
-                          userToken={session?.accessToken}/>
+        <AdvertForm dataObject={session?.user}
+                    userToken={session?.accessToken}
+                    successMessage='Заявка создана'/>
         </div>
     );
 }
